# src/modules/financial-goal/application/schema/financial-goal.schema.ts
import z from 'zod';
import { ERR_COMMON_EMPTY_PAYLOAD } from '@core/errors/messages/common.error';
import { AuditableSchema } from '@core/schema/auditable.schema';
import { IdSchema } from '@core/schema/common.schema';
import { createEntityQuerySchema } from '@core/factories/query.factory';

export const FinancialGoalSchema = z.object({
    id: IdSchema,
    userId: IdSchema,
    name: z.string().min(1).max(255),
    targetDate: z.string(),
    isActive: z.boolean().default(false),
    ...AuditableSchema.shape,
});

export const FinancialGoalCreateSchema = FinancialGoalSchema.pick({ name: true, targetDate: true });

export const FinancialGoalUpdateSchema = FinancialGoalCreateSchema.partial().refine(
    (data) => Object.keys(data).length > 0,
    {
        message: ERR_COMMON_EMPTY_PAYLOAD,
    },
);

export const visibleFinancialGoalColumns = [
    'id',
    'userId',
    'name',
    'targetDate',
    'isActive',
    'createdAt',
    'updatedAt',
] as const;

export const FinancialGoalQuerySchema = createEntityQuerySchema(
    visibleFinancialGoalColumns,
    FinancialGoalSchema.pick({ name: true, userId: true, isActive: true, targetDate: true }).shape,
);